+++ earthreader/web/osx.py
@@ -17,7 +17,7 @@ from six.moves import urllib, tkinter as tk
 from libearth.session import Session
 from waitress.server import create_server
 
-from . import app
+from . import create_app
 
 
 def open_webbrowser(port):
@@ -38,8 +38,7 @@ def main():
     directory = os.path.expanduser('~/.earthreader')
     repository = urllib.parse.urljoin('file://', directory)
     session_id = Session().identifier
-    app.config.update(REPOSITORY=repository, SESSION_ID=session_id,
-                      USE_WORKER=True)
+    app = create_app(REPOSITORY=repository, SESSION_ID=session_id)
     server = create_server(app, port=0)
     port = server.effective_port
     proc = threading.Thread(target=server.run)
@@ -47,7 +46,3 @@ def main():
     proc.start()
     open_webbrowser(port)
     root.mainloop()
-
-
-if __name__ == "__main__":
-    main()
