+++ earthreader/web/templates/base.html
@@ -8,7 +8,7 @@
 </head>
 <body>
 <h1>
-    <a href="{{ url_for('home') }}">Earth Reader</a>
+    <a href="{{ url_for('.home') }}">Earth Reader</a>
 </h1>
 {% block body %}
 {% endblock %}
