+++ earthreader/web/templates/category.html
@@ -4,7 +4,7 @@
     <ul>
         {% for feed in category.children %}
             <li>
-                <a href="{{ url_for('entries', feed_id=feed.feed_id) }}">{{ feed.label }}</a>
+                <a href="{{ url_for('.entries', feed_id=feed.feed_id) }}">{{ feed.label }}</a>
             </li>
         {% endfor %}
     </ul>
