+++ earthreader/web/templates/feed.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 {% block body %}
     <h2>
-        <a href="{{ url_for('entries', feed_id=feed_id) }}">{{ feed }}</a>
+        <a href="{{ url_for('.entries', feed_id=feed_id) }}">{{ feed }}</a>
     </h2>
     {% if feed.subtitle %}
         <p>{{ feed.subtitle }}</p>
